Backup 3. cannot detect bullet collision with enemy
for some reason. if i put code on the bullet and detect hitting the player, nothing works

but if i put code on the enemy and detect hitting the bullet, it works fine.

need to fix this at some point, but for now, it's 3AM and i need to be up at 9. i'mm settle for no bugs